requirements.txt
openai 
langchain
langchain-openai
langchain_pinecone
langchain[docarray]
docarray
pydantic==1.10.8
pytube 
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff
jupyterlab
streamlit
faiss-cpu
whisper
requests
bs4
numpy
pandas
pytube
pyPDF2
translate
unstructured
dotenv
langchain-community
scikit-learn
git+https://github.com/openai/whisper.git
